:recycle: Fixes #873 -- merged new command with existing command
Now all the new features and existing feature of sending an e-mail
are available.
