+++ bin/docker_start.sh
@@ -42,7 +42,7 @@ fi
 # specify username by setting OPENZAAK_SUPERUSER_USERNAME in the env
 # specify email by setting OPENZAAK_SUPERUSER_EMAIL in the env
 if [ -z "${OPENZAAK_SUPERUSER_USERNAME}" ]; then
-    python src/manage.py create_initial_superuser \
+    python src/manage.py createinitialsuperuser \
         --no-input \
         --username "${OPENZAAK_SUPERUSER_USERNAME}" \
         --email "${OPENZAAK_SUPERUSER_EMAIL:-admin\@admin.org}"
